requirements.txt
pytchat==0.5.5
python-dotenv==1.0.1
pyttsx3==2.90
pyvts==0.3.2
